succeded the flow v1 register until the end of purchase

+++ MadMicro.Service.OrderAPI/Models/DTO/OrderHeaderDTO.cs
@@ -6,8 +6,8 @@ public class OrderHeaderDTO
 
     public string? UserId { get; set; }
     public string? CouponCode { get; set; }
-    public double Discount { get; set; }
-    public double OrderTotal { get; set; }
+    public decimal Discount { get; set; }
+    public decimal OrderTotal { get; set; }
     public string? Name { get; set; }
     public string? Phone { get; set; }
     public string? Email { get; set; }
